requirements.txt
# Please use python V 3.7 to be compatible with all packages
gpytorch==1.5.0
torch==1.9.0
pyyaml==5.4.1
seaborn==0.11.2
scikit-learn==0.24.2
xgboost==1.4.0
tqdm==4.62.1
numpy==1.21.2
openml==0.12.2
catboost==0.26.1
auto-sklearn==0.14.5
hyperopt==0.2.5
configspace==0.4.21
lightgbm==3.3.5
tabpfn
ipdb
# autogluon==0.4.0
